MP1-4155: changed caching of processed data fixed bad response to specific requests from parsers or filters fixed seeking by time fixed rare freeze on stopping filter conceptually finished MPEG2 TS parser (it's still don't doing anything - except aligning stream)

+++ DirectShowFilters/MPUrlSourceSplitter/MPUrlSourceSplitter/MPUrlSourceSplitter_Protocol_Rtsp/RtspStreamTrack.cpp
@@ -99,7 +99,7 @@ int64_t CRtspStreamTrack::GetBytePosition(void)
     CRtspStreamFragment *firstFragment = this->GetStreamFragments()->GetItem(first);
     CRtspStreamFragment *lastFragment = this->GetStreamFragments()->GetItem(first + count - 1);
 
-    result = (unsigned int)(lastFragment->GetFragmentStartPosition() + lastFragment->GetLength() - firstFragment->GetFragmentStartPosition());
+    result = lastFragment->GetFragmentStartPosition() + (int64_t)lastFragment->GetLength() - firstFragment->GetFragmentStartPosition();
   }
 
   return result;
